deps/v8/test/webkit/fast/regex/malformed-escapes-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

This page tests handling of malformed escape sequences.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Testing regexp: /\ug/gm
PASS regexp.test('ug') is true
PASS regexp.lastIndex is 2

Testing regexp: /\xg/gm
PASS regexp.test('xg') is true
PASS regexp.lastIndex is 2

Testing regexp: /\c_/gm
PASS regexp.test('\\c_') is true
PASS regexp.lastIndex is 3

Testing regexp: /[\B]/gm
PASS regexp.test('B') is true
PASS regexp.lastIndex is 1

Testing regexp: /[\b]/gm
PASS regexp.test('\b') is true
PASS regexp.lastIndex is 1

Testing regexp: /\8/gm
PASS regexp.test('\\8') is true
PASS regexp.lastIndex is 2

Testing regexp: /^[\c]$/
PASS regexp.test('c') is true

Testing regexp: /^[\c_]$/
PASS regexp.test('c') is false

Testing regexp: /^[\c]]$/
PASS regexp.test('c]') is true
PASS successfullyParsed is true

TEST COMPLETE
